remove status service from add and edit page, fix unit tests
