+++ ConcernsCaseWork/ConcernsCaseWork.Redis/Models/CreateFinancialPlanModel.cs
@@ -9,7 +9,6 @@ namespace ConcernsCaseWork.Redis.Models
 		public DateTime CreatedAt { get; set; }
 		public DateTime UpdatedAt { get; set; }
 		public DateTime? DatePlanRequested { get; set; }
-		public DateTime? DateViablePlanReceived { get; set; }
 		public long? StatusId { get; set; }
 		public string Notes { get; set; }
 		public string CreatedBy { get; set; }
