+++ ConcernsCaseWork/ConcernsCaseWork.Tests/Pages/Case/Management/Action/FinancialPlan/AddPageModelTests.cs
@@ -56,17 +56,6 @@ namespace ConcernsCaseWork.Tests.Pages.Case.Management.Action.FinancialPlan
 				Times.Once);
 			
 			Assert.IsInstanceOf<PageResult>(response);
-			Assert.IsNotNull(pageModel.FinancialPlanStatuses);
-			Assert.AreEqual(2, pageModel.FinancialPlanStatuses.Count());
-			Assert.IsFalse(pageModel.FinancialPlanStatuses.Any(s => s.IsChecked));
-
-			var testStatus1 = pageModel.FinancialPlanStatuses.First();
-			Assert.IsTrue(validStatuses.Select(s => s.Description).Contains(testStatus1.Text));
-			Assert.IsTrue(validStatuses.Select(s => s.Name).Contains(testStatus1.Id));
-			
-			var testStatus2 = pageModel.FinancialPlanStatuses.Last();
-			Assert.IsTrue(validStatuses.Select(s => s.Description).Contains(testStatus2.Text));
-			Assert.IsTrue(validStatuses.Select(s => s.Name).Contains(testStatus2.Id));
 		}
 
 		[Test]
@@ -124,70 +113,6 @@ namespace ConcernsCaseWork.Tests.Pages.Case.Management.Action.FinancialPlan
 			Assert.That(pageModel.TempData["FinancialPlan.Message"], Is.EqualTo("Plan requested 00-00-0000 is an invalid date"));
 		}
 
-		[Test]
-		public async Task WhenOnPostAsync_Invalid_ViablePlanReceivedDate_ThrowsException_ReturnsPage()
-		{
-			// arrange
-			var mockFinancialPlanModelService = new Mock<IFinancialPlanModelService>();
-			var mockFinancialPlanStatusService = new Mock<IFinancialPlanStatusCachedService>();
-			var mockLogger = new Mock<ILogger<AddPageModel>>();
-
-			var statuses = FinancialPlanStatusFactory.BuildListOpenFinancialPlanStatusDto();
-			mockFinancialPlanStatusService.Setup(s => s.GetOpenFinancialPlansStatusesAsync()).ReturnsAsync(statuses);
-			
-			var pageModel = SetupAddPageModel(mockFinancialPlanModelService.Object, mockFinancialPlanStatusService.Object, mockLogger.Object);
-
-			var routeData = pageModel.RouteData.Values;
-			routeData.Add("urn", 1);
-
-			pageModel.HttpContext.Request.Form = new FormCollection(
-				new Dictionary<string, StringValues>
-				{
-					{ "dtr-day-viable-plan", new StringValues("00") },
-					{ "dtr-month-viable-plan", new StringValues("00") },
-					{ "dtr-year-viable-plan", new StringValues("0000") },
-				});
-
-			// act
-			var pageResponse = await pageModel.OnPostAsync();
-
-			// assert
-			Assert.That(pageResponse, Is.Not.Null);
-			Assert.That(pageModel.TempData, Is.Not.Null);
-			Assert.That(pageModel.TempData["FinancialPlan.Message"], Is.EqualTo("Viable plan 00-00-0000 is an invalid date"));
-		}
-		[Test]
-		public async Task WhenOnPostAsync_PartialViablePlanDate_ThrowsException_ReturnsPage()
-		{
-			// arrange
-			var mockFinancialPlanModelService = new Mock<IFinancialPlanModelService>();
-			var mockFinancialPlanStatusService = new Mock<IFinancialPlanStatusCachedService>();
-			var mockLogger = new Mock<ILogger<AddPageModel>>();
-			
-			var statuses = FinancialPlanStatusFactory.BuildListOpenFinancialPlanStatusDto();
-			mockFinancialPlanStatusService.Setup(s => s.GetOpenFinancialPlansStatusesAsync()).ReturnsAsync(statuses);
-			
-			var pageModel = SetupAddPageModel(mockFinancialPlanModelService.Object, mockFinancialPlanStatusService.Object, mockLogger.Object);
-
-			var routeData = pageModel.RouteData.Values;
-			routeData.Add("urn", 1);
-
-			pageModel.HttpContext.Request.Form = new FormCollection(
-				new Dictionary<string, StringValues>
-				{
-					{ "dtr-day-viable-plan", new StringValues("02") },
-					{ "dtr-month-viable-plan", new StringValues("04") },
-				});
-
-			// act
-			var pageResponse = await pageModel.OnPostAsync();
-
-			// assert
-			Assert.That(pageResponse, Is.Not.Null);
-			Assert.That(pageModel.TempData, Is.Not.Null);
-			Assert.That(pageModel.TempData["FinancialPlan.Message"], Is.EqualTo("Viable plan 02-04- is an invalid date"));
-		}
-		
 		[Test]
 		public async Task WhenOnPostAsync_Partial_RequestedDate_ThrowsException_ReturnsPage()
 		{
@@ -255,8 +180,7 @@ namespace ConcernsCaseWork.Tests.Pages.Case.Management.Action.FinancialPlan
 
 			// assert
 			mockFinancialPlanModelService.Verify(f => f.PostFinancialPlanByCaseUrn(It.Is<CreateFinancialPlanModel>(fpm =>
-					fpm.DatePlanRequested == new DateTime(year, month, day) &&
-					fpm.DateViablePlanReceived == null)), Times.Once);
+					fpm.DatePlanRequested == new DateTime(year, month, day))), Times.Once);
 			
 			Assert.IsNotNull(pageResponse);
 			Assert.IsNull(pageModel.TempData["Error.Message"]);
@@ -298,8 +222,7 @@ namespace ConcernsCaseWork.Tests.Pages.Case.Management.Action.FinancialPlan
 			// assert
 			mockFinancialPlanModelService.Verify(f => 
 				f.PostFinancialPlanByCaseUrn(It.Is<CreateFinancialPlanModel>(fpm => 
-						fpm.DateViablePlanReceived == new DateTime(year, month, day) &&
-				        fpm.DatePlanRequested == null)), Times.Once);
+						fpm.DatePlanRequested == null)), Times.Once);
 			
 			Assert.IsNotNull(pageResponse);
 			Assert.IsNull(pageModel.TempData["Error.Message"]);
@@ -355,7 +278,7 @@ namespace ConcernsCaseWork.Tests.Pages.Case.Management.Action.FinancialPlan
 		{
 			(PageContext pageContext, TempDataDictionary tempData, ActionContext actionContext) = PageContextFactory.PageContextBuilder(isAuthenticated);
 
-			var result = new AddPageModel(mockFinancialPlanModelService, mockFinancialPlanStatusService, mockLogger)
+			var result = new AddPageModel(mockFinancialPlanModelService, mockLogger)
 			{
 				PageContext = pageContext,
 				TempData = tempData,
