+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/FinancialPlan/Close.cshtml.cs
@@ -8,6 +8,9 @@ using ConcernsCaseWork.Models.CaseActions;
 using ConcernsCaseWork.Redis.FinancialPlan;
 using ConcernsCaseWork.Services.FinancialPlan;
 using ConcernsCaseWork.Service.FinancialPlan;
+using ConcernsCaseWork.Mappers;
+using System.Linq;
+using ConcernsCaseWork.Models;
 
 namespace ConcernsCaseWork.Pages.Case.Management.Action.FinancialPlan
 {
@@ -111,7 +114,34 @@ namespace ConcernsCaseWork.Pages.Case.Management.Action.FinancialPlan
 			return Page();
 		}
 		
-		protected override async Task<IList<FinancialPlanStatusDto>> GetAvailableStatusesAsync()
-			=> await _financialPlanStatusCachedService.GetClosureFinancialPlansStatusesAsync();
+
+
+		protected async Task<IList<FinancialPlanStatusDto>> GetAvailableStatusesAsync()
+		{
+			return await _financialPlanStatusCachedService.GetClosureFinancialPlansStatusesAsync();
+		}
+
+		protected virtual async Task<FinancialPlanStatusModel> GetOptionalStatusByNameAsync(string statusName)
+		{
+			var status = (await GetAvailableStatusesAsync())
+				.FirstOrDefault(s => s.Name.Equals(statusName));
+
+			return status is null ? null : FinancialPlanStatusMapping.MapDtoToModel(status);
+		}
+
+		protected async Task<IEnumerable<RadioItem>> GetStatusOptionsAsync(string selectedStatusName = null)
+			=> (await GetAvailableStatusesAsync())
+				.Select(s => new RadioItem
+				{
+					Id = s.Name,
+					Text = s.Description,
+					IsChecked = selectedStatusName == s.Name
+				});
+
+		protected async Task<FinancialPlanStatusModel> GetRequiredStatusByNameAsync(string statusName)
+		{
+			var status = await GetOptionalStatusByNameAsync(statusName);
+			return status ?? throw new InvalidOperationException($"Please select a reason for closing the Financial Plan");
+		}
 	}
 }
