+++ ConcernsCaseWork/ConcernsCaseWork/Pages/Case/Management/Action/FinancialPlan/Closed.cshtml
@@ -34,7 +34,6 @@
             </div>
         </h1>
 
-
         <div class="govuk-hint" id="financial-plan-hint">
             <a target="_blank" href="https://educationgovuk.sharepoint.com/sites/lveefa00003/SitePages/Trust%20financial%20plans.aspx" class="govuk-link" rel="noreferrer noopener">
                 SFSO Knowledge: Financial Plan
